requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
pyannote.audio>=3.1.0
openai-whisper>=20230314
yt-dlp>=2023.7.6
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
requests>=2.31.0
ollama>=0.1.7
webrtcvad>=2.0.10
soundfile>=0.12.1
librosa>=0.10.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
jupyter>=1.0.0
ipykernel>=6.0.0
tqdm>=4.65.0